meiduo_mis_sz37/meiduo_mall/meiduo_mall/apps/meiduo_admin/serializers/group_serializers.py
from django.contrib.auth.models import Group
from rest_framework import serializers

class GroupModeSerializer(serializers.ModelSerializer):

    class Meta:
        model = Group
        fields = '__all__'