meiduo_mis_sz37/meiduo_mall/requirements.txt
amqp==2.6.0
billiard==3.6.3.0
celery==4.4.5
certifi==2020.6.20
chardet==3.0.4
Django==2.2.5
django-cors-headers==3.3.0
django-crontab==0.7.1
django-haystack==2.8.1
django-redis==4.11.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
elasticsearch==2.4.1
fdfs-client-py==1.2.6
future==0.18.2
idna==2.7
importlib-metadata==1.6.1
itsdangerous==1.1.0
kombu==4.6.10
mutagen==1.41.0
mysqlclient==1.4.6
Pillow==7.1.2
pycryptodomex==3.9.8
PyJWT==1.7.1
python-alipay-sdk==1.8.0
pytz==2020.1
qiniu==7.2.8
QQLoginTool==0.3.0
redis==3.5.2
requests==2.19.1
sqlparse==0.3.1
urllib3==1.23
uWSGI==2.0.19.1
vine==1.3.0
zipp==3.1.0
